moved delta calculation out of websockets.js
