+++ lib/data.js
@@ -1,6 +1,9 @@
 'use strict';
 
+var _ = require('lodash');
 var async = require('async');
+var utils = require('./utils')();
+var ObjectID = require('mongodb').ObjectID;
 
 function uniq(a) {
   var seen = {};
@@ -42,6 +45,16 @@ function init (env, ctx) {
     return dir2Char[direction] || '-';
   }
 
+  data.clone = function clone ( ) {
+    return _.cloneDeep(data, function (value) {
+      //special handling of mongo ObjectID's
+      //see https://github.com/lodash/lodash/issues/602#issuecomment-47414964
+      if (value instanceof ObjectID) {
+        return value.toString();
+      }
+    });
+  };
+
   data.update = function update (done) {
 
     console.log('running data.update');
@@ -58,7 +71,6 @@ function init (env, ctx) {
 
     async.parallel({
       entries: function (callback) {
-        var now = new Date();
         var q = { find: {"date": {"$gte": earliest_data}} };
         ctx.entries.list(q, function (err, results) {
           if (!err && results) {
@@ -150,6 +162,91 @@ function init (env, ctx) {
     }, done);
 
   };
+
+  data.calculateDelta = function calculateDelta (lastData) {
+
+    var delta = {'delta': true};
+    var changesFound = false;
+
+    // if there's no updates done so far, just return the full set
+    if (!lastData.sgvs) return data;
+
+    console.log('lastData.sgvs last record time', lastData.sgvs[lastData.sgvs.length-1].x);
+    console.log('d.sgvslast record time', data.sgvs[data.sgvs.length-1].x);
+
+    function nsArrayDiff(oldArray, newArray) {
+      var seen = {};
+      var l = oldArray.length;
+      for (var i = 0; i < l; i++) { seen[oldArray[i].x] = true }
+      var result = [];
+      l = newArray.length;
+      for (var j = 0; j < l; j++) { if (!seen.hasOwnProperty(newArray[j].x)) { result.push(newArray[j]); console.log('delta data found'); } }
+      return result;
+    }
+
+    function sort (values) {
+      values.sort(function sorter (a, b) {
+        return a.x - b.x;
+      });
+    }
+
+    var sgvDelta = nsArrayDiff(lastData.sgvs, data.sgvs);
+
+    if (sgvDelta.length > 0) {
+      console.log('sgv changes found');
+      changesFound = true;
+      sort(sgvDelta);
+      delta.sgvs = sgvDelta;
+    }
+
+    var treatmentDelta = nsArrayDiff(lastData.treatments, data.treatments);
+
+    if (treatmentDelta.length > 0) {
+      console.log('treatment changes found');
+      changesFound = true;
+      sort(treatmentDelta);
+      delta.treatments = treatmentDelta;
+    }
+
+    var mbgsDelta = nsArrayDiff(lastData.mbgs, data.mbgs);
+
+    if (mbgsDelta.length > 0) {
+      console.log('mbgs changes found');
+      changesFound = true;
+      sort(mbgsDelta);
+      delta.mbgs = mbgsDelta;
+    }
+
+    var calsDelta = nsArrayDiff(lastData.cals, data.cals);
+
+    if (calsDelta.length > 0) {
+      console.log('cals changes found');
+      changesFound = true;
+      sort(calsDelta);
+      delta.cals = calsDelta;
+    }
+
+    if (JSON.stringify(lastData.devicestatus) != JSON.stringify(data.devicestatus)) {
+      console.log('devicestatus changes found');
+      changesFound = true;
+      delta.devicestatus = data.devicestatus;
+    }
+
+    if (JSON.stringify(lastData.profiles) != JSON.stringify(data.profiles)) {
+      console.log('profile changes found');
+      changesFound = true;
+      delta.profiles = data.profiles;
+    }
+
+    if (changesFound) {
+      console.log('changes found');
+      return delta;
+    }
+
+    return data;
+
+  };
+
 
   return data;
