+++ lib/utils.js
@@ -6,11 +6,11 @@ function init() {
     return utils;
   }
 
-    var MINUTE_IN_SECS = 60
-      , HOUR_IN_SECS = 3600
-      , DAY_IN_SECS = 86400;
+  var MINUTE_IN_SECS = 60
+    , HOUR_IN_SECS = 3600
+    , DAY_IN_SECS = 86400;
 
-    utils.toFixed = function toFixed(value) {
+  utils.toFixed = function toFixed(value) {
     if (value === 0) {
       return '0';
     } else {
@@ -47,8 +47,7 @@ function init() {
 
     return parts;
 
-  }
-
+  };
 
   return utils();
 }
