+++ lib/websocket.js
@@ -1,23 +1,7 @@
 'use strict';
 
 var _ = require('lodash');
-var ObjectID = require('mongodb').ObjectID;
-
-function nsArrayDiff(oldArray, newArray) {
-  var seen = {};
-  var l = oldArray.length;
-  for (var i = 0; i < l; i++) { seen[oldArray[i].x] = true }
-  var result = [];
-  l = newArray.length;
-  for (var j = 0; j < l; j++) { if (!seen.hasOwnProperty(newArray[j].x)) { result.push(newArray[j]); console.log('delta data found'); } }
-  return result;
-}
-
-function sort (values) {
-  values.sort(function sorter (a, b) {
-    return a.x - b.x;
-  });
-}
+var utils = require('./utils')();
 
 function init (env, ctx, server) {
 
@@ -25,12 +9,9 @@ function init (env, ctx, server) {
     return websocket;
   }
 
-  var lastUpdated = 0;
-  var patientData = {};
-  var patientDataUpdate = {};
-
   var io;
   var watchers = 0;
+  var lastData = {};
 
   function start ( ) {
     io = require('socket.io')({
@@ -43,18 +24,17 @@ function init (env, ctx, server) {
     });
   }
 
-  function emitData ( ) {
-    if (patientData.cals) {
-      console.log('running websocket.emitData', lastUpdated, patientDataUpdate.recentsgvs && patientDataUpdate.sgvdataupdate.length);
-      io.emit('dataUpdate', patientDataUpdate);
+  function emitData (delta) {
+    if (lastData.cals) {
+      console.log('running websocket.emitData', ctx.data.lastUpdated, delta.recentsgvs && delta.sgvdataupdate.length);
+      io.emit('dataUpdate', delta);
     }
   }
 
   function listeners ( ) {
     io.sockets.on('connection', function (socket) {
       // send all data upon new connection
-      socket.emit('dataUpdate',patientData);
-
+      socket.emit('dataUpdate',lastData);
       io.emit('clients', ++watchers);
       socket.on('ack', function(alarmType, silenceTime) {
         ctx.notifications.ack(alarmType, silenceTime);
@@ -70,32 +50,18 @@ function init (env, ctx, server) {
   }
 
   websocket.processData = function processData ( ) {
-
-    var d = ctx.data;
-    lastUpdated = d.lastUpdated;
-
     console.log('running websocket.processData');
-
-    var lastSGV = d.sgvs.length > 0 ? d.sgvs[d.sgvs.length - 1].y : null;
-
+    var lastSGV = ctx.data.sgvs.length > 0 ? ctx.data.sgvs[ctx.data.sgvs.length - 1].y : null;
     if (lastSGV) {
-      if (patientData.sgvs) {
-        var delta = calculateDelta(d);
+      if (lastData.sgvs) {
+        var delta = ctx.data.calculateDelta(lastData);
         if (delta.delta) {
-          patientDataUpdate = delta;
-          console.log('patientData full size', JSON.stringify(patientData).length,'bytes');
-          if (delta.sgvs) console.log('patientData update size', JSON.stringify(patientDataUpdate).length,'bytes');
-          emitData();
+          console.log('lastData full size', JSON.stringify(lastData).length,'bytes');
+          if (delta.sgvs) console.log('patientData update size', JSON.stringify(delta).length,'bytes');
+          emitData(delta);
         } else { console.log('delta calculation indicates no new data is present'); }
       }
-
-      //see https://github.com/lodash/lodash/issues/602#issuecomment-47414964
-      patientData = _.cloneDeep(d, function (value) {
-        if (value instanceof ObjectID) {
-          return value.toString();
-        }
-      });
-
+      lastData = ctx.data.clone();
     }
   };
 
@@ -108,73 +74,6 @@ function init (env, ctx, server) {
       console.info('emitted ' + info.type + ' to all clients');
     }
   };
-
-  function calculateDelta(d) {
-
-    var delta = {'delta': true};
-    var changesFound = false;
-
-    // if there's no updates done so far, just return the full set
-    if (!patientData.sgvs) return d;
-
-    console.log('patientData.sgvs last record time', patientData.sgvs[patientData.sgvs.length-1].x);
-    console.log('d.sgvslast record time', d.sgvs[d.sgvs.length-1].x);
-
-    var sgvDelta = nsArrayDiff(patientData.sgvs,d.sgvs);
-
-    if (sgvDelta.length > 0) {
-      console.log('sgv changes found');
-      changesFound = true;
-      sort(sgvDelta);
-      delta.sgvs = sgvDelta;
-    }
-
-    var treatmentDelta = nsArrayDiff(patientData.treatments,d.treatments);
-
-    if (treatmentDelta.length > 0) {
-      console.log('treatment changes found');
-      changesFound = true;
-      sort(treatmentDelta);
-      delta.treatments = treatmentDelta;
-    }
-
-    var mbgsDelta = nsArrayDiff(patientData.mbgs,d.mbgs);
-
-    if (mbgsDelta.length > 0) {
-      console.log('mbgs changes found');
-      changesFound = true;
-      sort(mbgsDelta);
-      delta.mbgs = mbgsDelta;
-    }
-
-    var calsDelta = nsArrayDiff(patientData.cals,d.cals);
-
-    if (calsDelta.length > 0) {
-      console.log('cals changes found');
-      changesFound = true;
-      sort(calsDelta);
-      delta.cals = calsDelta;
-    }
-
-    if (JSON.stringify(patientData.devicestatus) != JSON.stringify(d.devicestatus)) {
-      console.log('devicestatus changes found');
-      changesFound = true;
-      delta.devicestatus = d.devicestatus;
-    }
-
-    if (JSON.stringify(patientData.profiles) != JSON.stringify(d.profiles)) {
-      console.log('profile changes found');
-      changesFound = true;
-      delta.profiles = d.profiles;
-    }
-
-    if (changesFound) {
-      console.log('changes found');
-      return delta;
-    }
-    return d;
-
-  }
 
   start( );
   listeners( );
